+++ NEWS.md
@@ -2,6 +2,7 @@
   - Examples updated
   - dutility function no longer casts warnings when handling formulas
   - Efficient estimation of recurrent events mean: recurrentMarginalAIPCW 
+  - Average treatment effect for competing risks: binregATE 
 
 # mets 1.2.8.1
   - Maintenance release